changed all math functions to std to use the c++ functions rather than the C functions which deal with integral values only. Fixed some bugs with QSMatrix functionality which caused bugs for LLL. Fixed ILLL bugs. Implemented a new calculation of the Dirichlet Coeff

+++ CPPTestBench/main.cpp
@@ -289,7 +289,7 @@ void IteratedLLLTest() {
 		preValues = GetILLLRandomizedMatrix(m, n);
 		std::cout << "Input Matrix:\n" << preValues << std::endl;
 
-		std::vector< QSMatrix<int> > result = IteratedLLL_Dyadic(preValues, alpha, epsilon, qmax, M);
+		std::vector< QSMatrix<double> > result = IteratedLLL_Dyadic(preValues, alpha, epsilon, qmax, M);
 		for (size_t k = 0; k < result.size(); k++) {
 			std::cout << "Result " << k + 1 << ":\n" << result[k] << std::endl;
 		}
